Fix midi loop issue and dangling devices with rtpmidid
Everytime the playback device selection was used a new RtMidi endpoint
was created.  Using the connect all after this was causing completely
erattic behavior such as loops and lag.  The old endpoints must be
deleted before creating new endpoints.

The connectall() should also restore the input and output endpoints
for the light bar.

Signed-off-by: Jason Wessel <jason.wessel@windriver.com>