Add the ability to use an alternate font directory
When running the visualizer on a system without X installed and on a
readonly file system, there are no fonts.  This commit makes it
possible to use an alternate location from a git checkout.

It also fails early if the fonts cannot be found vs waiting for it to
fail when the screensaver runs.

Signed-off-by: Jason Wessel <jason.wessel@windriver.com>